test(mp-service): Refactor data model tests into discrete files.
