+++ test/Core.Test/Auth/UserFeatures/UserMasterPassword/MasterPasswordServiceTests.cs
@@ -1013,366 +1013,4 @@ public class MasterPasswordServiceTests
         }
     }
 
-    // --- Data model validation: SetInitialPasswordData ---
-
-    public class SetInitialPasswordDataTests
-    {
-        private static User BuildValidSetInitialUser()
-        {
-            var user = new User
-            {
-                Id = Guid.NewGuid(),
-                Email = "test@example.com",
-                MasterPassword = null,
-                Key = null,
-                MasterPasswordSalt = null,
-                UsesKeyConnector = false,
-                Kdf = KdfType.PBKDF2_SHA256,
-                KdfIterations = 600000
-            };
-            return user;
-        }
-
-        private static SetInitialPasswordData BuildData(User user, string? saltOverride = null)
-        {
-            // Stage 1: salt == email while MasterPasswordSalt is null (PM-28143 separates them in Stage 3).
-            var salt = saltOverride ?? user.GetMasterPasswordSalt();
-            var kdf = new KdfSettings
-            {
-                KdfType = user.Kdf,
-                Iterations = user.KdfIterations,
-                Memory = user.KdfMemory,
-                Parallelism = user.KdfParallelism
-            };
-            return new SetInitialPasswordData
-            {
-                MasterPasswordUnlock = new MasterPasswordUnlockData
-                {
-                    Salt = salt,
-                    MasterKeyWrappedUserKey = "wrapped-key",
-                    Kdf = kdf
-                },
-                MasterPasswordAuthentication = new MasterPasswordAuthenticationData
-                {
-                    Salt = salt,
-                    MasterPasswordAuthenticationHash = "hash",
-                    Kdf = kdf
-                }
-            };
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Accepts_WhenUserHasNoMasterPassword()
-        {
-            var user = BuildValidSetInitialUser();
-            var data = BuildData(user);
-
-            // Should not throw
-            data.ValidateDataForUser(user);
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserHasMasterPassword()
-        {
-            var user = BuildValidSetInitialUser();
-            user.MasterPassword = "existing-hash";
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserHasKey()
-        {
-            var user = BuildValidSetInitialUser();
-            user.Key = "existing-key";
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserHasSalt()
-        {
-            var user = BuildValidSetInitialUser();
-            user.MasterPasswordSalt = "existing-salt";
-            var data = BuildData(user, saltOverride: "existing-salt");
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserIsKeyConnector()
-        {
-            var user = BuildValidSetInitialUser();
-            user.UsesKeyConnector = true;
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenSaltMismatch()
-        {
-            var user = BuildValidSetInitialUser();
-            var data = BuildData(user, saltOverride: "wrong-salt");
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenAuthenticationSaltMismatch_UnlockSaltCorrect()
-        {
-            var user = BuildValidSetInitialUser();
-            var correctSalt = user.GetMasterPasswordSalt();
-            var kdf = new KdfSettings
-            {
-                KdfType = user.Kdf,
-                Iterations = user.KdfIterations,
-                Memory = user.KdfMemory,
-                Parallelism = user.KdfParallelism
-            };
-            // Authentication salt is wrong; Unlock salt is correct.
-            var data = new SetInitialPasswordData
-            {
-                MasterPasswordUnlock = new MasterPasswordUnlockData
-                {
-                    Salt = correctSalt,
-                    MasterKeyWrappedUserKey = "wrapped-key",
-                    Kdf = kdf
-                },
-                MasterPasswordAuthentication = new MasterPasswordAuthenticationData
-                {
-                    Salt = "wrong-auth-salt",
-                    MasterPasswordAuthenticationHash = "hash",
-                    Kdf = kdf
-                }
-            };
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUnlockSaltMismatch_AuthenticationSaltCorrect()
-        {
-            var user = BuildValidSetInitialUser();
-            var correctSalt = user.GetMasterPasswordSalt();
-            var kdf = new KdfSettings
-            {
-                KdfType = user.Kdf,
-                Iterations = user.KdfIterations,
-                Memory = user.KdfMemory,
-                Parallelism = user.KdfParallelism
-            };
-            // Unlock salt is wrong; Authentication salt is correct.
-            var data = new SetInitialPasswordData
-            {
-                MasterPasswordUnlock = new MasterPasswordUnlockData
-                {
-                    Salt = "wrong-unlock-salt",
-                    MasterKeyWrappedUserKey = "wrapped-key",
-                    Kdf = kdf
-                },
-                MasterPasswordAuthentication = new MasterPasswordAuthenticationData
-                {
-                    Salt = correctSalt,
-                    MasterPasswordAuthenticationHash = "hash",
-                    Kdf = kdf
-                }
-            };
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-    }
-
-    // --- Data model validation: UpdateExistingPasswordData ---
-
-    public class UpdateExistingPasswordDataTests
-    {
-        private static User BuildValidUpdateUser()
-        {
-            var user = new User
-            {
-                Id = Guid.NewGuid(),
-                Email = "test@example.com",
-                MasterPassword = "existing-hash",
-                Key = "existing-key",
-                MasterPasswordSalt = "stored-salt",
-                UsesKeyConnector = false,
-                Kdf = KdfType.PBKDF2_SHA256,
-                KdfIterations = 600000
-            };
-            return user;
-        }
-
-        private static UpdateExistingPasswordData BuildData(User user, string? saltOverride = null,
-            KdfSettings? kdfOverride = null)
-        {
-            var salt = saltOverride ?? user.GetMasterPasswordSalt();
-            var kdf = kdfOverride ?? new KdfSettings
-            {
-                KdfType = user.Kdf,
-                Iterations = user.KdfIterations,
-                Memory = user.KdfMemory,
-                Parallelism = user.KdfParallelism
-            };
-            return new UpdateExistingPasswordData
-            {
-                MasterPasswordUnlock = new MasterPasswordUnlockData
-                {
-                    Salt = salt,
-                    MasterKeyWrappedUserKey = "wrapped-key",
-                    Kdf = kdf
-                },
-                MasterPasswordAuthentication = new MasterPasswordAuthenticationData
-                {
-                    Salt = salt,
-                    MasterPasswordAuthenticationHash = "hash",
-                    Kdf = kdf
-                }
-            };
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Accepts_WhenUserHasMasterPassword_KdfAndSaltMatch()
-        {
-            var user = BuildValidUpdateUser();
-            var data = BuildData(user);
-
-            // Should not throw
-            data.ValidateDataForUser(user);
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserHasNoMasterPassword()
-        {
-            var user = BuildValidUpdateUser();
-            user.MasterPassword = null;
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserIsKeyConnector()
-        {
-            var user = BuildValidUpdateUser();
-            user.UsesKeyConnector = true;
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenKdfChanged()
-        {
-            var user = BuildValidUpdateUser();
-            var mismatchedKdf = new KdfSettings
-            {
-                KdfType = KdfType.Argon2id,
-                Iterations = 3,
-                Memory = 64,
-                Parallelism = 4
-            };
-            var data = BuildData(user, kdfOverride: mismatchedKdf);
-
-            Assert.Throws<ArgumentException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenSaltChanged()
-        {
-            var user = BuildValidUpdateUser();
-            var data = BuildData(user, saltOverride: "wrong-salt");
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-    }
-
-    // --- Data model validation: UpdateExistingPasswordAndKdfData ---
-
-    public class UpdateExistingPasswordAndKdfDataTests
-    {
-        private static User BuildValidUser()
-        {
-            return new User
-            {
-                Id = Guid.NewGuid(),
-                Email = "test@example.com",
-                MasterPassword = "existing-hash",
-                Key = "existing-key",
-                MasterPasswordSalt = "stored-salt",
-                UsesKeyConnector = false,
-                Kdf = KdfType.PBKDF2_SHA256,
-                KdfIterations = 600000
-            };
-        }
-
-        private static UpdateExistingPasswordAndKdfData BuildData(User user, string? saltOverride = null)
-        {
-            var salt = saltOverride ?? user.GetMasterPasswordSalt();
-            var newKdf = new KdfSettings
-            {
-                KdfType = KdfType.Argon2id,
-                Iterations = 3,
-                Memory = 64,
-                Parallelism = 4
-            };
-            return new UpdateExistingPasswordAndKdfData
-            {
-                MasterPasswordUnlock = new MasterPasswordUnlockData
-                {
-                    Salt = salt,
-                    MasterKeyWrappedUserKey = "wrapped-key",
-                    Kdf = newKdf
-                },
-                MasterPasswordAuthentication = new MasterPasswordAuthenticationData
-                {
-                    Salt = salt,
-                    MasterPasswordAuthenticationHash = "hash",
-                    Kdf = newKdf
-                }
-            };
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Accepts_WhenUserHasMasterPassword_SaltMatch_KdfChanged()
-        {
-            var user = BuildValidUser();
-            var data = BuildData(user);
-
-            // Should not throw — KDF change is permitted here
-            data.ValidateDataForUser(user);
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserHasNoMasterPassword()
-        {
-            var user = BuildValidUser();
-            user.MasterPassword = null;
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenUserIsKeyConnector()
-        {
-            var user = BuildValidUser();
-            user.UsesKeyConnector = true;
-            var data = BuildData(user);
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-
-        [Fact]
-        public void ValidateDataForUser_Throws_WhenSaltChanged()
-        {
-            var user = BuildValidUser();
-            var data = BuildData(user, saltOverride: "wrong-salt");
-
-            Assert.Throws<BadRequestException>(() => data.ValidateDataForUser(user));
-        }
-    }
-
 }
